add seperate cont and pres components
